\name{print.summary.tri}
\title{Print a summary of a triangulation object}
\author{A. Gebhardt}
\usage{\method{print}{summary.tri}(x, ...)}
\alias{print.summary.tri}
\arguments{
  \item{x}{object of class \code{"summary.tri"},
    generated by \code{\link{summary.tri}}.}
  \item{...}{additional paramters for \code{print}}
}
\description{
  Prints some information about \code{tri.obj}
}
\value{None
}
\references{
  R. J. Renka (1996). Algorithm 751: TRIPACK: a constrained
  two-dimensional Delaunay triangulation package.
  ACM Transactions on Mathematical Software.
  \bold{22}, 1-8.
}
\seealso{
  \code{\link{tri}},\code{\link{tri.mesh}},
  \code{\link{print.tri}}, \code{\link{plot.tri}},
  \code{\link{summary.tri}}. 
}
\keyword{spatial}
