\name{print.summary.voronoi}
\title{Print a summary of a voronoi object}
\author{A. Gebhardt}
\usage{print.summary.voronoi(voronoi.summ.obj)}
print(summary(voronoi.obj))
}
\alias{print.summary.voronoi}
\arguments{
  \item{voronoi.summ.obj}{object of class \code{"summary.voronoi"},
    generated by \code{\link{summary.voronoi}}.}
}
\description{
  Prints some information about \code{voronoi.obj}
}
\value{None
}
\references{
  R. J. Renka (1996). Algorithm 751: TRIPACK: a constrained
  two-dimensional {Delaunay} triangulation package.
  ACM Transactions on Mathematical Software.
  \bold{22}, 1-8.
}
\seealso{
  \code{\link{voronoi}},\code{\link{voronoi.mosaic}},
  \code{\link{print.voronoi}}, \code{\link{plot.voronoi}},
  \code{\link{summary.voronoi}}. 
}
\keyword{}
