\name{print.tri}
\title{Print a triangulation object}
\author{A. Gebhardt}
\usage{\method{print}{tri}(x,...)
}
\alias{print.tri}
\arguments{
  \item{x}{object of class \code{"tri"}}
  \item{...}{additional paramters for \code{print}}
}
\description{prints a adjacency list of \code{"x"}
}
\value{None
}
\references{
    R. J. Renka (1996). Algorithm 751: TRIPACK: a constrained
  two-dimensional {Delaunay} triangulation package.
  ACM Transactions on Mathematical Software.
  \bold{22}, 1-8.
}
\seealso{
  \code{\link{tri}},
  \code{\link{plot.tri}},
  \code{\link{summary.tri}}
}
\keyword{spatial}
