\name{print.tri}
\title{Print a voronoi object}
\author{A. Gebhardt}
\usage{print(voronoi.obj)
}
\alias{print.voronoi}
\arguments{
  \item{voronoi.obj}{object of class \code{"voronoi"}}
}
\description{prints a summary of \code{"voronoi.obj"}
}
\value{None
}
\references{
    R. J. Renka (1996). Algorithm 751: TRIPACK: a constrained
  two-dimensional {Delaunay} triangulation package.
  ACM Transactions on Mathematical Software.
  \bold{22}, 1-8.
}
\seealso{
  \code{\link{voronoi}},
  \code{\link{plot.voronoi}},
  \code{\link{summary.voronoi}}
}
\keyword{}
