\name{tritest}
\alias{tritest}
\non_function{}
\title{ tritest / sample data }
\description{
A very simply set set of points to test the tripack functions, taken
from the FORTRAN original.
}
\references{
  R. J. Renka (1996). Algorithm 751: TRIPACK: a constrained
  two-dimensional {Delaunay} triangulation package.
  ACM Transactions on Mathematical Software.
  \bold{22}, 1-8.
}
\keywords{data}
