\name{print.tri}
\title{Print a triangulation object}
\author{A. Gebhardt}
\usage{print(tri.obj)
}
\alias{print.tri}
\arguments{
  \item{tri.obj}{object of class \code{"tri"}}
}
\description{prints a adjacency list of \code{"tri.obj"}
}
\value{None
}
\references{
    R. J. Renka (1996). Algorithm 751: TRIPACK: a constrained
  two-dimensional {Delaunay} triangulation package.
  ACM Transactions on Mathematical Software.
  \bold{22}, 1-8.
}
\seealso{
  \code{\link{tri}},
  \code{\link{plot.tri}},
  \code{\link{summary.tri}}
}
\keyword{spatial}
