\name{print.voronoi}
\title{Print a xect}
\author{A. Gebhardt}
\usage{\method{print}{voronoi}(x,...)
}
\alias{print.voronoi}
\arguments{
  \item{x}{object of class \code{"voronoi"}}
  \item{...}{additional paramters for \code{print}}
}
\description{prints a summary of \code{"x"}
}
\value{None
}
\references{
    R. J. Renka (1996). Algorithm 751: TRIPACK: a constrained
  two-dimensional {Delaunay} triangulation package.
  ACM Transactions on Mathematical Software.
  \bold{22}, 1-8.
}
\seealso{
  \code{\link{voronoi}},
  \code{\link{plot.voronoi}},
  \code{\link{summary.voronoi}}
}
\keyword{spatial}
