\name{summary.voronoi}
\title{Return a summary of a voronoi object}
\author{A. Gebhardt}
\usage{summary(voronoi.obj)
}
\alias{summary.voronoi}
\arguments{
  \item{voronoi.obj}{object of class \code{"voronoi"}}
}
\description{
  Returns some information about \code{voronoi.obj}
}
\value{Object of class \code{"summary.voronoi"}.

  It contains the number of nodes (\code{nn}) and dummy nodes (\code{nd}).
}
\references{
    R. J. Renka (1996). Algorithm 751: TRIPACK: a constrained
  two-dimensional {Delaunay} triangulation package.
  ACM Transactions on Mathematical Software.
  \bold{22}, 1-8.
}
\seealso{
  \code{\link{voronoi}},\code{\link{voronoi.mosaic}},
  \code{\link{print.voronoi}}, \code{\link{plot.voronoi}},
  \code{\link{print.summary.voronoi}}. 
}
<<<<<<< summary.voronoi.Rd
\keyword{spatial}
=======
\keyword{spatial}
>>>>>>> 1.4
