\name{tripack-internal}
\title{Internal functions}
\alias{voronoi.polyarea}
\alias{tri.vordist}
\alias{voronoi.findvertices}
\alias{tri.swap}
\alias{tri.swap.lowlevel}
\description{
 Internal tripack functions
}
\details{
  These functions are not intended to be called by the user.
 
}
\keyword{spatial}
